\name{fdrtool-internal}

\alias{pvt.isoMean}
\alias{r2t}
\alias{t2r}

\alias{pvt.nullfunction}
\alias{pvt.plotlabels}
\alias{ecdf.pval}

\alias{num.curv}

\alias{approximate.fit}
\alias{iqr.fit}

\alias{get.nullmodel}
\alias{pvt.fit.nullmodel}




\title{Internal fdrtool Functions}
\description{
  Internal fdrtool functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
